modulization & major improvements

+++ handlers_pishu/video.php
@@ -1,12 +1,10 @@
 <?php
-//on video sent
-if ($update->message->video->file_size > 5000000) {
-    $telegram->sendMessage($user_id, "لطفا کلیپی با حجم کمتر از ۵ مگابایت ارسال کنید");
-} else {
-    echo "Video sent\n";
 
-    $background = $user->read("telegram_user_id", $user_id)['background'];
-    if(file_exists($background)) {
+if ($update->message->video->file_size > 5000000) {
+    dm("لطفا کلیپی با حجم کمتر از ۵ مگابایت ارسال کنید");
+} else {
+    $background = $user_record['background'];
+    if($background && file_exists($background)) {
         unlink($background);
     }
 
@@ -14,10 +12,12 @@ if ($update->message->video->file_size > 5000000) {
     $result = $telegram->getFile($file_id);
     $file_path = json_decode($result)->result->file_path;
 
-    if($telegram->download($user_id, $file_path) == False) {
-        $telegram->sendMessage($user_id, "تنظیم ویدیوی پس‌زمینه موفقیت آمیز نبود، دوباره امتحان کنید");
+    $file_name = $telegram->download($user_id, $file_path);
+    if(!$file_name) {
+        dm("تنظیم ویدیوی پس‌زمینه موفقیت آمیز نبود، دوباره امتحان کنید");
     } else {
-        $telegram->sendMessage($user_id, "ویدیوی پس‌زمینه تنظیم شد");
-        $user->update("checkpoint", "", "telegram_user_id", $user_id);
+        dm("ویدیوی پس‌زمینه تنظیم شد");
+        setCheckpoint("background");
+        userUpdate("background", $file_name);
     }
 }
